Changes required by the code reviewer
