+++ src/Pages/Mission/Mission.css
@@ -31,10 +31,6 @@ table tbody tr:nth-child(even) {
   background-color: #f2f2f2;
 }
 
-.member {
-  white-space: nowrap;
-}
-
 .not {
   background-color: #838181;
   color: #fff;
@@ -51,7 +47,7 @@ table tbody tr:nth-child(even) {
   font-size: 0.8rem;
 }
 
-.btn-join {
+.btn-join, .member {
   white-space: nowrap;
 }
