+++ src/Pages/rockets/Rockets.css
@@ -3,7 +3,7 @@
   justify-content: space-between;
   margin: 1rem auto;
   width: 90%;
-  flex-wrap: wrap;
+  /* flex-wrap: wrap; */
 }
 
 .rocket-img img {
