+++ src/__test__/MyProfile.test.js
@@ -2,16 +2,16 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../../redux/store';
-import Navbar from './Navbar';
+import store from '../redux/store';
+import MyProfile from '../Pages/MyProfile/MyProfile';
 
-it('Navbar should render', () => {
-  const navbar = render(
+it('Profile should render', () => {
+  const profile = render(
     <Provider store={store}>
       <BrowserRouter>
-        <Navbar />
+        <MyProfile />
       </BrowserRouter>
     </Provider>,
   );
-  expect(navbar).toMatchSnapshot();
+  expect(profile).toMatchSnapshot();
 });
