+++ src/__test__/Navbar.test.js
@@ -2,16 +2,16 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../../redux/store';
-import MyProfile from './MyProfile';
+import store from '../redux/store';
+import Navbar from '../components/Navbar/Navbar';
 
-it('Profile should render', () => {
-  const profile = render(
+it('Navbar should render', () => {
+  const navbar = render(
     <Provider store={store}>
       <BrowserRouter>
-        <MyProfile />
+        <Navbar />
       </BrowserRouter>
     </Provider>,
   );
-  expect(profile).toMatchSnapshot();
+  expect(navbar).toMatchSnapshot();
 });
