+++ src/__test__/Rockets.test.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../../redux/store';
-import Rockets from './Rockets';
+import store from '../redux/store';
+import Rockets from '../Pages/rockets/Rockets';
 
 it('Rockets should render', () => {
   const rockets = render(
